Exception handling - delete

+++ src/main/java/com/ddiniz/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.ddiniz.course.entities.User;
 import com.ddiniz.course.repositories.UserRepository;
+import com.ddiniz.course.services.exceptions.DatabaseException;
 import com.ddiniz.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -28,9 +30,16 @@ public class UserService {
 	public User insert(User obj) {
 		return repository.save(obj);
 	}
-	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    }catch (DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
+	} 
 	public User update(Long id, User obj) {
 		//User entity = repository.getReferenceById(id);
 		User entity = findById(id);
